Finalized PowerBI Dashboards for mini Mag7 project.
Much of the project was focused on familiarizing myself with the software, hence the very basic colour schemes and general aesthetics. Definitely not because colours are my weak point. 

Dashboards feature slicers to target date ranges and companies, very basic data visualisation, and toggleables to move between boards or to show/hide data.